@@ -9,13 +9,13 @@
     <version>1.0.0</version>
 
     <properties>
-        <java.version>11</java.version>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <java.version>17</java.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
 
         <libre.version>1.0.0</libre.version>
-        <springboot.version>2.7.8</springboot.version>
-        <libre.dependencies.version>2.7.8</libre.dependencies.version>
+        <springboot.version>3.1.1</springboot.version>
+        <libre.dependencies.version>3.1.1</libre.dependencies.version>
 
         <bitwalker.version>1.21</bitwalker.version>
         <oshi.version>6.2.2</oshi.version>
