@@ -35,6 +35,7 @@ import org.springframework.security.oauth2.server.authorization.web.authenticati
 import org.springframework.security.web.DefaultSecurityFilterChain;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.AuthenticationConverter;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.security.web.util.matcher.RequestMatcher;
 
 import java.security.KeyPair;
@@ -70,18 +71,19 @@ public class AuthorizationServerConfiguration {
 				.authorizationEndpoint(authorizationEndpoint -> authorizationEndpoint// 授权码端点个性化confirm页面
 						.consentPage(SecurityConstants.CUSTOM_CONSENT_PAGE_URI)));
 
-		RequestMatcher endpointsMatcher = authorizationServerConfigurer.getEndpointsMatcher();
-		DefaultSecurityFilterChain securityFilterChain = http.requestMatcher(endpointsMatcher)
-				.authorizeRequests(authorizeRequests -> authorizeRequests.anyRequest().authenticated())
+		AntPathRequestMatcher[] requestMatchers = new AntPathRequestMatcher[] {
+				AntPathRequestMatcher.antMatcher("/token/**"), AntPathRequestMatcher.antMatcher("/actuator/**"),
+				AntPathRequestMatcher.antMatcher("/css/**"), AntPathRequestMatcher.antMatcher("/error") };
 
-				.apply(authorizationServerConfigurer.authorizationService(authorizationService)// redis存储token的实现
-						.authorizationServerSettings(
-								AuthorizationServerSettings.builder()
-										.issuer("https://www.baidu.com")
-										.tokenEndpoint("/api/oauth2/token")
-										.build()))
-				// 授权码登录的登录页个性化
-				.and().apply(new FormIdentityLoginConfigurer()).and().build();
+		http.authorizeHttpRequests(authorizeRequests -> {
+			// 自定义接口、端点暴露
+			authorizeRequests.requestMatchers(requestMatchers).permitAll();
+			authorizeRequests.anyRequest().authenticated();
+		}).apply(authorizationServerConfigurer.authorizationService(authorizationService)// redis存储token的实现
+				.authorizationServerSettings(AuthorizationServerSettings.builder().issuer("https://www.libre.com")
+						.tokenEndpoint("/api/oauth2/token").build()));
+		http.apply(new FormIdentityLoginConfigurer());
+		DefaultSecurityFilterChain securityFilterChain = http.build();
 
 		// 注入自定义授权模式实现
 		addCustomOAuth2GrantAuthenticationProvider(http);
@@ -145,5 +147,5 @@ public class AuthorizationServerConfiguration {
 		// 处理 OAuth2ResourceOwnerSmsAuthenticationToken
 		http.authenticationProvider(resourceOwnerSmsAuthenticationProvider);
 	}
-}
 
+}
