@@ -6,6 +6,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
+import org.springframework.security.config.annotation.web.configurers.RequestCacheConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
 
 /**
@@ -25,9 +28,10 @@ public class WebSecurityConfiguration {
 	 */
 	@Bean
 	SecurityFilterChain defaultSecurityFilterChain(HttpSecurity http) throws Exception {
-		http.authorizeRequests(authorizeRequests -> authorizeRequests.antMatchers("/token/*").permitAll()// 开放自定义的部分端点
-				.anyRequest().authenticated()).headers().frameOptions().sameOrigin()// 避免iframe同源无法登录
-				.and().apply(new FormIdentityLoginConfigurer()); // 表单登录个性化
+		http.authorizeHttpRequests(authorizeRequests -> authorizeRequests.requestMatchers("/token/*").permitAll()// 开放自定义的部分端点
+				.anyRequest().authenticated())
+				.headers(h -> h.frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin))// 避免iframe同源无法登录
+				.apply(new FormIdentityLoginConfigurer()); // 表单登录个性化
 		// 处理 UsernamePasswordAuthenticationToken
 		http.authenticationProvider(new OAuth2DaoAuthenticationProvider());
 		return http.build();
@@ -44,9 +48,10 @@ public class WebSecurityConfiguration {
 	@Bean
 	@Order(0)
 	SecurityFilterChain resources(HttpSecurity http) throws Exception {
-		http.requestMatchers((matchers) -> matchers.antMatchers("/actuator/**", "/css/**", "/error"))
-				.authorizeHttpRequests((authorize) -> authorize.anyRequest().permitAll()).requestCache().disable()
-				.securityContext().disable().sessionManagement().disable();
+		http.authorizeHttpRequests((matchers) -> matchers.requestMatchers("/actuator/**", "/css/**", "/error"))
+				.authorizeHttpRequests((authorize) -> authorize.anyRequest().permitAll())
+				.requestCache(RequestCacheConfigurer::disable).securityContext(AbstractHttpConfigurer::disable)
+				.sessionManagement(AbstractHttpConfigurer::disable);
 		return http.build();
 	}
 
